+++ editor/include/adminer.inc.php
@@ -69,6 +69,10 @@ class Adminer {
 		return $csp;
 	}
 
+	function verifyVersion() {
+		return true;
+	}
+
 	function head($dark = null) {
 		return true;
 	}
@@ -572,7 +576,7 @@ ORDER BY ORDINAL_POSITION", null, "") as $row
 	function navigation($missing) {
 		echo "<h1>" . adminer()->name() . " <span class='version'>" . VERSION;
 		$new_version = $_COOKIE["adminer_version"];
-		echo " <a href='https://www.adminer.org/editor/#download'" . target_blank() . " id='version'>" . (version_compare(VERSION, $new_version) < 0 ? h($new_version) : "") . "</a>";
+		echo " <a href='https://www.adminer.org/editor/#download'" . target_blank() . " id='version'>" . (version_compare(VERSION, $new_version) < 0 ? h($new_version) : "") . version_iframe() . "</a>";
 		echo "</span></h1>\n";
 		switch_lang();
 		if ($missing == "auth") {